Anteultimos cambios predemo

+++ db/schema.rb
@@ -89,7 +89,7 @@ ActiveRecord::Schema[7.0].define(version: 2022_11_25_201433) do
   end
 
   create_table "wallets", force: :cascade do |t|
-    t.float "money", default: 0.0
+    t.integer "money", default: 0
     t.integer "user_id", null: false
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
